requirements.txt
Flask
gunicorn
scikit-learn
numpy
pandas
PyYAML
evidently
matplotlib
dill
xlrd
-e .